[DSPDC-1710] DAP terraform (#140)
Why

We need to capture our current DAP infrastructure in Terraform.

This PR

Creates a "base" environment at environments/dap, with concrete instantations of that environment in subdirs dev/ and prod/. The intent is to avoid our current pattern of copying+pasting the world around. The DAP base is also not generic enough to warrant placement in our templates dir, hence this approach.